iTXt chunk processed
