+++ Project1/main.py
@@ -18,6 +18,7 @@ def choose_photo():
     chunk_processor.gAMA_chunk_processor()
     chunk_processor.cHRM_chunk_processor()
     chunk_processor.tEXt_chunk_processor()
+    chunk_processor.iTXt_chunk_processor()
     chunk_processor.IEND_chunk_processor()
     query = chunk_processor.return_chunks_names_query()
     display_photo(chunk_processor)
